files added

+++ src/Vehicle.java
@@ -1,7 +1,11 @@
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Types;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Scanner;
 
 public class Vehicle {
@@ -14,7 +18,9 @@ public class Vehicle {
             System.out.println("2. View all vehicle information");
             System.out.println("3. Update vehicle information");
             System.out.println("4. Delete vehicle information");
-            System.out.println("5. Return to the home screen");
+            System.out.println("5. Assign Vehicle to permit");
+            System.out.println("6. Remove Vehicle from permit");
+            System.out.println("7. Return to the home screen");
 
             int choice;
             while (true) {
@@ -32,7 +38,9 @@ public class Vehicle {
                 case 2 -> viewVehicle();
                 case 3 -> updateVehicle();
                 case 4 -> deleteVehicle();
-                case 5 -> {
+                case 5 -> assignVehicleToPermit();
+                case 6 -> removeVehicleFromPermit();
+                case 7 -> {
                     System.out.println("<-- Back to home menu");
                     exit = true;
                 }
@@ -54,8 +62,6 @@ public class Vehicle {
         String color = scanner.nextLine();
         System.out.println("Enter year of vehicle: ");
         String year = scanner.nextLine();
-//        System.out.println("Enter permit ID of vehicle (Enter N/A if no permit ID): ");
-//        String permitID = scanner.nextLine();
 
         if (doesLicenseNoExist(licenseNo)) {
             System.out.println("License Number already exists. Please try again.");
@@ -68,7 +74,8 @@ public class Vehicle {
     }
 
     public static void viewVehicle() throws SQLException {
-        Main.printTable("Vehicle");
+        ResultSet rs = Main.statement.executeQuery("SELECT LicenseNo, Model, Manf, Color, Year FROM Vehicle;");
+        DBTablePrinter.printResultSet(rs);
     }
 
     private static void deleteVehicle() throws SQLException {
@@ -76,8 +83,28 @@ public class Vehicle {
         String licenseNo = scanner.nextLine();
 
         if (doesLicenseNoExist(licenseNo)) {
-            Main.statement.executeUpdate("DELETE FROM Vehicle WHERE LicenseNo = \'" + licenseNo + "\';");
-            System.out.println("Vehicle row with license number " + licenseNo + " deleted successfully.");
+            List<String> cNumbers = Citation.getCitationNumberFromLicense(licenseNo);
+            if (cNumbers != null) {
+                for (String value : cNumbers) {
+                    if (Citation.doesCitationNoExist(value)) {
+                        String paymentStatus = Citation.getColumnDetails("PaymentStatus", value);
+                        String appealStatus = Citation.getColumnDetails("AppealStatus", value);
+                        if (appealStatus.equalsIgnoreCase("APPROVED") || paymentStatus.equalsIgnoreCase("PAID")) {
+                            Main.statement.executeUpdate("DELETE FROM Encompasses WHERE CNumber = \'" + value + "\';");
+                            Main.statement.executeUpdate("DELETE FROM Citation WHERE CNumber = \'" + value + "\';");
+                            Main.statement.executeUpdate("DELETE FROM Vehicle WHERE LicenseNo = \'" + licenseNo + "\';");
+                            System.out.println("Vehicle row with license number " + licenseNo + " deleted successfully.");
+                        }else{
+                            System.out.println("Cannot remove the vehicle due to unpaid citations.");
+                        }
+                    } else {
+                        System.out.println("Invalid citation number entered. Please try again."); //remove
+                    }
+                }
+            } else {
+                Main.statement.executeUpdate("DELETE FROM Vehicle WHERE LicenseNo = \'" + licenseNo + "\';");
+                System.out.println("Vehicle row with license number " + licenseNo + " deleted successfully.");
+            }
         } else {
             System.out.println("Incorrect license number entered. Please select from the below license numbers:");
             ResultSet ids = Main.statement.executeQuery("SELECT LicenseNo FROM Vehicle;");
@@ -182,10 +209,125 @@ public class Vehicle {
 
     public static boolean doesLicenseNoExist(String licenseNo) throws SQLException {
         boolean vehicleExists = false;
-        ResultSet rs = Main.statement.executeQuery("SELECT * FROM Vehicle WHERE LicenseNo = \'" + licenseNo+"\';");
+        ResultSet rs = Main.statement.executeQuery("SELECT * FROM Vehicle WHERE LicenseNo = \'" + licenseNo + "\';");
         if (rs.next()) {
             vehicleExists = true;
         }
         return vehicleExists;
     }
+
+    public static void assignVehicleToPermit() throws SQLException {
+        String permitId, licenseNo;
+        do {
+            System.out.println("Enter Permit ID: ");
+            permitId = scanner.nextLine();
+            if (!Permit.doesPermitIDExist(permitId))
+                System.out.println("Incorrect Permit ID entered. Please try again.");
+        } while (!Permit.doesPermitIDExist(permitId));
+        do {
+            System.out.println("Enter License Number: ");
+            licenseNo = scanner.nextLine();
+            if (!Vehicle.doesLicenseNoExist(licenseNo))
+                System.out.println("Incorrect license number entered. Please try again.");
+        } while (!Vehicle.doesLicenseNoExist(licenseNo));
+
+        assignVehicleToPermit(permitId, licenseNo);
+    }
+
+    public static void assignVehicleToPermit(String permitID, String licenseNo) throws SQLException {
+        int cs = 0;
+        String insertQuery = "UPDATE Vehicle SET PermitID = \'" + permitID + "\' WHERE LicenseNo = \'" + licenseNo + "\';";
+        String status = null;
+        BigInteger driverID = BigInteger.valueOf(0);
+        if (Permit.doesPermitIDExist(permitID)) {
+            if (doesLicenseNoExist(licenseNo)) {
+                ResultSet id = Main.statement.executeQuery("SELECT DriverID FROM Permit WHERE PermitID = \'" + permitID + "\';");
+                if (id.next()) {
+                    driverID = id.getBigDecimal("DriverID").toBigInteger();
+                } else {
+                    System.out.println("Driver ID not found for the given permit. Please try again.");
+                }
+                ResultSet rs = Main.statement.executeQuery("SELECT Status FROM Driver WHERE DriverID = " + driverID + ";");
+                if (rs.next()) {
+                    status = rs.getString("Status");
+                    String countQuery = "Select Count(PermitID) AS CountP from Vehicle WHERE DriverID = " + driverID;
+                    ResultSet cns = Main.statement.executeQuery(countQuery);
+                    if (cns.next()) {
+                        cs = cns.getInt("CountP");
+                        System.out.println(cs);
+                    }
+                }
+                HashMap<String, Integer> hashMap = new HashMap<String, Integer>();
+                hashMap.put("E", 2);
+                hashMap.put("S", 1);
+                hashMap.put("V", 1);
+                int expectedCount = hashMap.get(status);
+                if (cs < expectedCount) {
+                    Main.statement.executeUpdate(insertQuery);
+                    System.out.println("Vehicle " + licenseNo + " assigned to permit " + permitID + " successfully.");
+                } else {
+                    System.out.println(Permit.CANNOT_ASSIGN_PERMIT_MESSAGE);
+                }
+            } else {
+                System.out.println("Incorrect license number entered. Please select from the below license numbers:");
+                printAllLicenseNumbers();
+            }
+        } else {
+            System.out.println("Incorrect permit ID entered. Please select from the below permit IDs:");
+            Permit.printAllPermitIds();
+        }
+    }
+
+    public static void removeVehicleFromPermit() throws SQLException {
+        boolean flag = false;
+        String permitID, licenseNo;
+        do {
+            System.out.println("Enter Permit ID: ");
+            permitID = scanner.nextLine();
+            if (!Permit.doesPermitIDExist(permitID))
+                System.out.println("Incorrect Permit ID entered. Please try again.");
+        } while (!Permit.doesPermitIDExist(permitID));
+        do {
+            System.out.println("Enter License Number: ");
+            licenseNo = scanner.nextLine();
+            if (!Vehicle.doesLicenseNoExist(licenseNo))
+                System.out.println("Incorrect license number entered. Please try again.");
+        } while (!Vehicle.doesLicenseNoExist(licenseNo));
+
+        List<String> cNumbers = Citation.getCitationNumberFromLicense(licenseNo);
+        if (cNumbers != null) {// needs work
+            System.out.println("Cannot remove the vehicle due to following citations: ");
+            for (String value : cNumbers) {
+                System.out.println(value);
+            }
+        } else {
+            while (!flag) {
+                if (Permit.doesPermitIDExist(permitID)) {
+                    if (doesLicenseNoExist(licenseNo)) {
+                        flag = true;
+                        PreparedStatement ps = Main.connection.prepareStatement("UPDATE Vehicle SET PermitID = ? WHERE LicenseNo = \'" + licenseNo + "\';");
+                        ps.setNull(1, Types.VARCHAR);
+                        ps.executeUpdate();
+                        System.out.println("Vehicle " + licenseNo + " removed from permit " + permitID + " successfully.");
+                    } else {
+                        System.out.println("Incorrect license number entered. Please select from the below license numbers:");
+                        printAllLicenseNumbers();
+                    }
+                } else {
+                    System.out.println("Incorrect permit ID entered. Please select from the below permit IDs:");
+                    Permit.printAllPermitIds();
+                }
+            }
+        }
+    }
+
+    private static void printAllLicenseNumbers() throws SQLException {
+        ResultSet licenses = Main.statement.executeQuery("SELECT LicenseNo FROM Vehicle;");
+        while (licenses.next()) {
+            System.out.println(licenses.getString("LicenseNo"));
+        }
+        System.out.println();
+    }
+
+
 }
